Import database, configuration, core from upstream
- Database
An osquery database is basically just a SQLite3 database with several
virtual tables attached. This method is the main abstraction for creating
SQLite3 databases within osquery.

- Configuration
osquery has two types on configurations. Things that don't change during
the execution of the process should be configured as command-line
arguments. Things that can change during the lifetime of program execution
should be defined using the osquery::config::Config class and the pluggable
plugin interface that is included with it.

Signed-off-by: sangwan.kwon <sangwan.kwon@samsung.com>