Import scheduler from upstream
osquery comes with a scheduler, which schedules a variety of things. This
is one of the core parts of the osqueryd daemon. To use this, simply use
this function as your entry point when creating a new thread.

TBD: Fix sql_tests failed cases. (It should be come with build-in tables.)

Signed-off-by: sangwan.kwon <sangwan.kwon@samsung.com>